+++ src/client/deployStrategies/auraDeploy.ts
@@ -5,7 +5,6 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { readFileSync } from 'fs';
-import { AuraDefinition } from '../../utils/deploy';
 import { extName, baseName } from '../../utils';
 import { SourcePath } from '../../common';
 import {
@@ -13,6 +12,7 @@ import {
   ComponentDeployment,
   ToolingDeployStatus,
   SourceDeployResult,
+  AuraDefinition,
 } from '../types';
 import { deployTypes } from '../toolingApi';
 import { BaseDeploy } from './baseDeploy';
