+++ src/utils/fileSystemHandler.ts
@@ -22,25 +22,6 @@ export function ensureFileExists(filePath: string): void {
   fs.closeSync(fs.openSync(filePath, 'w'));
 }
 
-export function isDirectory(fsPath: SourcePath): boolean {
-  return fs.lstatSync(fsPath).isDirectory();
-}
-
-export function deleteDirectory(dirPath: string): void {
-  if (fs.existsSync(dirPath)) {
-    const files = fs.readdirSync(dirPath);
-    for (const file of files) {
-      const curPath = path.join(dirPath, file);
-      if (isDirectory(curPath)) {
-        deleteDirectory(curPath);
-      } else {
-        fs.unlinkSync(path.join(dirPath, file));
-      }
-    }
-    fs.rmdirSync(dirPath);
-  }
-}
-
 /**
  * Method to save multiple files on disk.
  *
